% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{myPaste}
\alias{myPaste}
\title{Paste of vector elements}
\usage{
myPaste(arg1, arg2)
}
\arguments{
\item{arg1}{character A character string.}

\item{arg2}{numeric A numerical value to append
to the character string provided in \code{arg1}.}
}
\value{
character A string with \code{arg1} and
\code{arg2} pasted together
}
\description{
\code{myPaste} returns the sum of all the values
present in its arguments.
}
\details{
This is a generic function: methods can be
defined for it directly or via the
\code{\link[=Summary]{Summary()}} group generic. For this to work
properly, the arguments \code{...} should be
unnamed, and dispatch is on the first
argument.
}
\examples{
myPaste("Jacques' cat is ", 3)
}
